test(inspect): extract shared DataGrid test fixtures and drag harness
The four DataGrid behavior test files re-declared identical row/column
fixtures and the jsdom drag-and-drop harness (dataTransfer stub,
dragHeader, headerOrder). Move them into a data-grid testFixtures module.
