+++ apps/inspect/src/app/shared/data-grid/DataGrid.pinning.test.tsx
@@ -1,25 +1,18 @@
-import {
-  act,
-  cleanup,
-  fireEvent,
-  render,
-  screen,
-} from "@testing-library/react";
+import { cleanup, render, screen } from "@testing-library/react";
 import { afterEach, describe, expect, test, vi } from "vitest";
 
-import type { ExtendedColumnDef } from "./columnTypes";
 import { DataGrid } from "./DataGrid";
+import type { AbcRow as Row } from "./testFixtures";
+import {
+  dragHeader,
+  headerOrder,
+  makeAbcColumns as makeColumns,
+  abcRows as rows,
+} from "./testFixtures";
 
 // Vitest globals aren't enabled in this app, so RTL's automatic afterEach
 // cleanup never fires. Run it explicitly.
 afterEach(cleanup);
-
-interface Row {
-  id: string;
-  a: string;
-  b: string;
-  c: string;
-}
 
 // closest() returns Element; the assertions below read HTMLElement members.
 const asHtmlElement = (node: Element | null): HTMLElement => {
@@ -28,47 +21,6 @@ const asHtmlElement = (node: Element | null): HTMLElement => {
   }
   return node;
 };
-
-const rows: Row[] = [{ id: "1", a: "1a", b: "1b", c: "1c" }];
-
-const makeColumns = (pinnedIds: string[]): ExtendedColumnDef<Row>[] =>
-  (["a", "b", "c"] as const).map((key) => ({
-    id: key,
-    header: key.toUpperCase(),
-    size: 100,
-    accessorFn: (r: Row) => r[key],
-    cell: ({ getValue }) => <div>{getValue<string>()}</div>,
-    ...(pinnedIds.includes(key) ? { pinned: "start" as const } : {}),
-  }));
-
-// jsdom has no DataTransfer; provide the bits the handlers touch.
-const dataTransfer = () => ({
-  effectAllowed: "",
-  dropEffect: "",
-  setData: () => {},
-  getData: () => "",
-});
-
-/** Drag the header labelled `from` and drop it on the header cell of `to`. */
-const dragHeader = async (from: string, to: string) => {
-  const dt = dataTransfer();
-  const source = screen.getByText(from);
-  const target = screen.getByText(to).closest('[role="columnheader"]');
-  expect(target).not.toBeNull();
-  if (!target) return;
-  fireEvent.dragStart(source, { dataTransfer: dt });
-  // The drag-source state flip is deferred a macrotask (see DataGrid's
-  // handleHeaderDragStart) — flush it before the drag proceeds.
-  await act(async () => {
-    await new Promise((resolve) => setTimeout(resolve, 0));
-  });
-  fireEvent.dragOver(target, { dataTransfer: dt });
-  fireEvent.drop(target, { dataTransfer: dt });
-  fireEvent.dragEnd(source, { dataTransfer: dt });
-};
-
-const headerOrder = () =>
-  screen.getAllByRole("columnheader").map((cell) => cell.textContent.trim());
 
 const headerCell = (label: string) =>
   asHtmlElement(screen.getByText(label).closest('[role="columnheader"]'));
