+++ apps/inspect/src/app/shared/data-grid/DataGrid.reorder.test.tsx
@@ -1,66 +1,20 @@
-import {
-  act,
-  cleanup,
-  fireEvent,
-  render,
-  screen,
-} from "@testing-library/react";
+import { cleanup, render } from "@testing-library/react";
 import { afterEach, describe, expect, test, vi } from "vitest";
 
-import type { ExtendedColumnDef } from "./columnTypes";
 import { DataGrid } from "./DataGrid";
+import type { AbcRow as Row } from "./testFixtures";
+import {
+  dragHeader,
+  headerOrder,
+  makeAbcColumns,
+  abcRows as rows,
+} from "./testFixtures";
 
 // Vitest globals aren't enabled in this app, so RTL's automatic afterEach
 // cleanup never fires. Run it explicitly.
 afterEach(cleanup);
 
-interface Row {
-  id: string;
-  a: string;
-  b: string;
-  c: string;
-}
-
-const rows: Row[] = [{ id: "1", a: "1a", b: "1b", c: "1c" }];
-
-const columns: ExtendedColumnDef<Row>[] = (["a", "b", "c"] as const).map(
-  (key) => ({
-    id: key,
-    header: key.toUpperCase(),
-    size: 100,
-    accessorFn: (r: Row) => r[key],
-    cell: ({ getValue }) => <div>{getValue<string>()}</div>,
-  })
-);
-
-// jsdom has no DataTransfer; provide the bits the handlers touch.
-const dataTransfer = () => ({
-  effectAllowed: "",
-  dropEffect: "",
-  setData: () => {},
-  getData: () => "",
-});
-
-/** Drag the header labelled `from` and drop it on the header cell of `to`. */
-const dragHeader = async (from: string, to: string) => {
-  const dt = dataTransfer();
-  const source = screen.getByText(from);
-  const target = screen.getByText(to).closest('[role="columnheader"]');
-  expect(target).not.toBeNull();
-  if (!target) return;
-  fireEvent.dragStart(source, { dataTransfer: dt });
-  // The drag-source state flip is deferred a macrotask (see DataGrid's
-  // handleHeaderDragStart) — flush it before the drag proceeds.
-  await act(async () => {
-    await new Promise((resolve) => setTimeout(resolve, 0));
-  });
-  fireEvent.dragOver(target, { dataTransfer: dt });
-  fireEvent.drop(target, { dataTransfer: dt });
-  fireEvent.dragEnd(source, { dataTransfer: dt });
-};
-
-const headerOrder = () =>
-  screen.getAllByRole("columnheader").map((cell) => cell.textContent.trim());
+const columns = makeAbcColumns();
 
 describe("DataGrid column reorder", () => {
   test("dropping a header on another column reorders and reports the order", async () => {
